refactor: 입력 관련 로직을 InputController로 분리 및 Application에 반영
- PrintController 클래스의 입력 로직을 InputController 클래스로 이동하여 역할 분리
- Application 클래스에 InputController를 사용하여 코드 리팩토링
