+++ src/main/java/lotto/Application.java
@@ -6,10 +6,14 @@ public class Application {
     public static void main(String[] args) {
 
         PrintController printController = new PrintController();
-        int purchaseNumber = printController.inputPurchaseNumber();
+        InputController inputController = new InputController();
+
+        int purchaseNumber = inputController.inputPurchaseNumber();
         List<Lotto> lottos = printController.printLottos(purchaseNumber);
-        List<Integer> winNumber = printController.inputWinNumber();
-        int bonusNumber = printController.inputBonusNumber(winNumber);
+
+        List<Integer> winNumber = inputController.inputWinNumber();
+        int bonusNumber = inputController.inputBonusNumber(winNumber);
+
         WinNumbers winNumbers = new WinNumbers(winNumber, bonusNumber);
         LottoResult lottoResult = printController.printStatistics(lottos, winNumbers);
         printController.printRateOfReturn(lottoResult,purchaseNumber);
